+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   * **added** serialport plugin, works in electron and Chrome browser
   * **added** update plugin, handles updates in electron and browser
   * **added** to docs Adafruit CLUE board setup
+  * **added** tutorials plugin
   * **fixed** schematics and enbled it for browser
   * **fixed** display micropython options after port selection in browser
   * **fixed** schematics image layout